requirements.txt
pyzmq
msgpack-python
nose
